pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spinscribe"
version = "0.1.0"
description = "Multi-Agent Content Creation System"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "camel-ai[all]>=0.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "redis>=5.0.0",
    "qdrant-client>=1.7.0",
    "asyncpg>=0.29.0",
]

